add policy for content and lesson class

+++ app/controllers/lessons_controller.rb
@@ -1,10 +1,10 @@
 class LessonsController < ApplicationController
-  rescue_from Pundit::NotAuthorizedError, with: :user_not_authorized
   layout 'dashboard'
   before_action :authenticate_user!
 
   def create_lesson
     @course = Course.find_by(id: params[:course_id])
+    authorize @course, policy_class: LessonPolicy
     @lesson = @course.lessons.build(lesson_params)
     if @lesson.save
       flash[:notice] = 'A new lesson is added'
@@ -17,6 +17,7 @@ class LessonsController < ApplicationController
 
   def destroy_lesson
     @lesson = Lesson.find_by(id: params[:lesson_id])
+    authorize @lesson
     course_id = @lesson.course_id
     if @lesson.destroy
       flash[:notice] = "This lesson is deleted"
@@ -30,10 +31,12 @@ class LessonsController < ApplicationController
   def edit_lesson
     @page_title = 'Dashboard -> Edit a lesson'
     @lesson = Lesson.find_by(id: params[:lesson_id])
+    authorize @lesson
   end
 
   def save_lesson
     @lesson = Lesson.find_by(id: params[:lesson_id])
+    authorize @lesson, :edit_lesson?
     if @lesson.update(lesson_params)
       flash[:notice] = "This lesson is updated"
       redirect_to "/dashboard/show_a_course/#{@lesson.course_id}"
@@ -45,15 +48,10 @@ class LessonsController < ApplicationController
 
   def show_lesson
     @course = Course.includes(:lessons).find_by(id: params[:course_id])
+    authorize @course, :show_course?, policy_class: CoursePolicy
   end
 
   private
-
-  def user_not_authorized
-    flash[:notice] = 'You are not authorized to perform this action.'
-    redirect_to dashboard_show_user_path
-  end
-
   def lesson_params
     params.require(:lesson).permit(:title, :description)
   end
